Reorganization
Sepeated the code for the dataStructure, to make it easier to mantain in the future.
