+++ Main/depend.txt
@@ -1,3 +1,3 @@
 main.cpp
-dataStructure.cpp
+dataStructure/dataStructure.cpp
 Entity.cpp
